% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{bounds_for_ci}
\alias{bounds_for_ci}
\title{Bounds for CI root finding}
\usage{
bounds_for_ci(
  y,
  ct,
  alpha,
  lower_ci_bound = TRUE,
  bound_inflation_factor = 0.05
)
}
\arguments{
\item{y}{Numeric. Observed statistic.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}

\item{lower_ci_bound}{Logical. Finding lower (TRUE) or upper (FALSE) CI bound.}

\item{bound_inflation_factor}{Numeric. Factor to inflate bounds. For numeric stability.}
}
\value{
Numeric vector of length 2: search interval.
}
\description{
Bounds for CI root finding
}
\keyword{internal}
