% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{calculate_ar_length}
\alias{calculate_ar_length}
\title{Calculate AR length}
\usage{
calculate_ar_length(ar)
}
\arguments{
\item{ar}{Numeric vector of length 4.}
}
\value{
Numeric. Total length.
}
\description{
Calculate AR length
}
\keyword{internal}
