% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{cond_ar_given_lb}
\alias{cond_ar_given_lb}
\title{Construct AR given lower bound}
\usage{
cond_ar_given_lb(lb, theta, ct, alpha)
}
\arguments{
\item{lb}{Numeric. Lower bound.}

\item{theta}{Numeric. Parameter value.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric vector of length 4.
}
\description{
Construct AR given lower bound
}
\keyword{internal}
