% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{cond_ar_given_lb_ub}
\alias{cond_ar_given_lb_ub}
\title{Construct AR given lower and upper bounds}
\usage{
cond_ar_given_lb_ub(lb, ub, ct)
}
\arguments{
\item{lb}{Numeric. Lower bound.}

\item{ub}{Numeric. Upper bound.}

\item{ct}{Numeric. Critical value.}
}
\value{
Numeric vector of length 4: c(ll, ul, lr, ur).
}
\description{
Construct AR given lower and upper bounds
}
\keyword{internal}
