% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{cond_ar_given_ub}
\alias{cond_ar_given_ub}
\title{Construct AR given upper bound}
\usage{
cond_ar_given_ub(ub, theta, ct, alpha)
}
\arguments{
\item{ub}{Numeric. Upper bound.}

\item{theta}{Numeric. Parameter value.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric vector of length 4.
}
\description{
Construct AR given upper bound
}
\keyword{internal}
