% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_normal.R
\name{density_tn_outer}
\alias{density_tn_outer}
\title{Density of N(mu, 1) truncated outside (-b, b)}
\usage{
density_tn_outer(x, mu, b)
}
\arguments{
\item{x}{Numeric vector. Points at which to evaluate the density.}

\item{mu}{Numeric. Mean of the underlying normal distribution.}

\item{b}{Numeric. Truncation bound.}
}
\value{
Numeric vector. Density values.
}
\description{
Density of N(mu, 1) truncated outside (-b, b)
}
