% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{diff_from_shortest_given_lb}
\alias{diff_from_shortest_given_lb}
\title{Difference from shortest AR given lower bound}
\usage{
diff_from_shortest_given_lb(target_len, lb, theta, ct, alpha)
}
\arguments{
\item{target_len}{Numeric. Target length.}

\item{lb}{Numeric. Lower bound.}

\item{theta}{Numeric. Parameter value.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric. Difference from target.
}
\description{
Difference from shortest AR given lower bound
}
\keyword{internal}
