% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{find_tilde_1}
\alias{find_tilde_1}
\title{Find tilde_1: value where min(AR(r)) > ct}
\usage{
find_tilde_1(ct, r, alpha)
}
\arguments{
\item{ct}{Numeric. Critical value.}

\item{r}{Numeric. Inflation factor.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric. The theta_tilde_1 value.
}
\description{
Find tilde_1: value where min(AR(r)) > ct
}
\keyword{internal}
