% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{invert_ars}
\alias{invert_ars}
\title{Invert acceptance regions to get CI}
\usage{
invert_ars(y, ars)
}
\arguments{
\item{y}{Numeric. Observed statistic.}

\item{ars}{Matrix. Acceptance regions.}
}
\value{
Numeric vector. CI bounds.
}
\description{
Invert acceptance regions to get CI
}
\keyword{internal}
