% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{lb_finder}
\alias{lb_finder}
\title{Find lower bound given upper bound}
\usage{
lb_finder(ub, theta, ct, alpha)
}
\arguments{
\item{ub}{Numeric. Upper bound.}

\item{theta}{Numeric. Parameter value.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric. Lower bound.
}
\description{
Find lower bound given upper bound
}
\keyword{internal}
