% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_normal.R
\name{quantile_tn_outer}
\alias{quantile_tn_outer}
\title{Quantile function for N(mu, 1) truncated outside (-b, b)}
\usage{
quantile_tn_outer(p, mu, b)
}
\arguments{
\item{p}{Numeric in [0, 1]. Probability.}

\item{mu}{Numeric. Mean of the underlying normal distribution.}

\item{b}{Numeric. Truncation bound.}
}
\value{
Numeric. Quantile value.
}
\description{
Quantile function for N(mu, 1) truncated outside (-b, b)
}
