% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditional.R
\name{ub_finder}
\alias{ub_finder}
\title{Find upper bound given lower bound}
\usage{
ub_finder(lb, theta, ct, alpha)
}
\arguments{
\item{lb}{Numeric. Lower bound.}

\item{theta}{Numeric. Parameter value.}

\item{ct}{Numeric. Critical value.}

\item{alpha}{Numeric. Significance level.}
}
\value{
Numeric. Upper bound.
}
\description{
Find upper bound given lower bound
}
\keyword{internal}
